src/requirements.txt
absl-py
albumentations==0.4.3
cachetools
certifi
chardet
cycler
decorator
future
google-auth
google-auth-oauthlib
grpcio
idna
imageio
imgaug
joblib
kiwisolver
Markdown
matplotlib
networkx
numpy==1.24.4
oauthlib
opencv-python
opencv-python-headless
Pillow
protobuf
pyasn1
pyasn1-modules
pyparsing
python-dateutil
PyWavelets
PyYAML
requests
requests-oauthlib
rsa
scikit-image
scikit-learn
scipy
six
tensorboard
--extra-index-url https://download.pytorch.org/whl/cu126
torch
torchvision
tqdm
urllib3
Werkzeug
pycocotools